+++ replacement/lru/lru.cc
@@ -2,7 +2,7 @@
 
 #include <algorithm>
 
-uint32_t lru::find_victim(uint32_t triggering_cpu, uint64_t instr_id, uint32_t set, const CACHE::BLOCK* current_set, uint64_t ip, uint64_t full_addr, uint32_t type)
+long lru::find_victim(uint32_t triggering_cpu, uint64_t instr_id, long set, const CACHE::BLOCK* current_set, uint64_t ip, uint64_t full_addr, uint32_t type)
 {
   auto begin = std::next(std::begin(last_used_cycles), set * NUM_WAY);
   auto end = std::next(begin, NUM_WAY);
@@ -11,10 +11,10 @@ uint32_t lru::find_victim(uint32_t triggering_cpu, uint64_t instr_id, uint32_t s
   auto victim = std::min_element(begin, end);
   assert(begin <= victim);
   assert(victim < end);
-  return static_cast<uint32_t>(std::distance(begin, victim)); // cast protected by prior asserts
+  return std::distance(begin, victim);
 }
 
-void lru::update_replacement_state(uint32_t triggering_cpu, uint32_t set, uint32_t way, uint64_t full_addr, uint64_t ip, uint64_t victim_addr, uint32_t type, uint8_t hit)
+void lru::update_replacement_state(uint32_t triggering_cpu, long set, long way, uint64_t full_addr, uint64_t ip, uint64_t victim_addr, uint32_t type, uint8_t hit)
 {
   // Mark the way as being used on the current cycle
   if (!hit || type != WRITE) // Skip this for writeback hits
