+++ test/cpp/src/432-prefetch-metadata.cc
@@ -24,7 +24,7 @@ struct metadata_collector : champsim::modules::prefetcher
     return metadata_in;
   }
 
-  uint32_t prefetcher_cache_fill(uint64_t, uint32_t, uint32_t, uint8_t, uint64_t, uint32_t metadata_in)
+  uint32_t prefetcher_cache_fill(uint64_t, long, long, uint8_t, uint64_t, uint32_t metadata_in)
   {
     auto it = test::metadata_fill_collector.try_emplace(intern_);
     it.first->second.push_back(metadata_in);
@@ -38,7 +38,7 @@ struct metadata_fill_emitter : champsim::modules::prefetcher
   using prefetcher::prefetcher;
 
   uint32_t prefetcher_cache_operate(uint64_t, uint64_t, uint8_t, uint8_t, uint32_t metadata_in) { return metadata_in; }
-  uint32_t prefetcher_cache_fill(uint64_t, uint32_t, uint32_t, uint8_t, uint64_t, uint32_t) { return to_emit; }
+  uint32_t prefetcher_cache_fill(uint64_t, long, long, uint8_t, uint64_t, uint32_t) { return to_emit; }
 };
 
 SCENARIO("Prefetch metadata from an issued prefetch is seen in the lower level") {
