+++ test/cpp/src/441-replacement-bypass.cc
@@ -8,7 +8,7 @@ template <uint64_t bypass_addr>
 struct bypass_replacement : champsim::modules::replacement
 {
   using replacement::replacement;
-  auto find_victim(uint32_t, uint64_t, uint32_t, const CACHE::BLOCK*, uint64_t, uint64_t addr, uint32_t)
+  auto find_victim(uint32_t, uint64_t, long, const CACHE::BLOCK*, uint64_t, uint64_t addr, uint32_t)
   {
     if (addr == bypass_addr)
       return 1;
